Move BoundaryHistory test to its own file
There is a relatively small amount of code duplication between the
History and BoundaryHistory tests.  Conforming to the usual file
structure is worth it.
