+++ tests/Unit/Time/CMakeLists.txt
@@ -4,6 +4,7 @@
 set(LIBRARY "Test_Time")
 
 set(LIBRARY_SOURCES
+  Test_BoundaryHistory.cpp
   Test_EvolutionOrdering.cpp
   Test_History.cpp
   Test_SelfStart.cpp
